Collate NukeStudio tasks extraction.

+++ pyblish_bumpybox/plugins/nukestudio/extract_nukestudio_tasks.py
@@ -11,38 +11,29 @@ class ExtractHieroNukeStudioTask(pyblish.api.InstancePlugin):
 
     def process(self, instance):
         import time
+        import os
+
+        import hiero.core.nuke as nuke
 
         task = instance[0]
         task.startTask()
         while task.taskStep():
             time.sleep(1)
 
+        families = instance.data["families"]
+        if "mov" in families or "img" in families:
+            script_path = instance[0]._scriptfile
+            log_path = script_path.replace(".nk", ".log")
+            log_file = open(log_path, "w")
+            process = nuke.executeNukeScript(script_path, log_file, True)
 
-class ExtractHierNukeStudioTranscode(pyblish.api.InstancePlugin):
-    """Extract transcode."""
+            self.poll(process)
 
-    order = ExtractHieroNukeStudioTask.order + 0.1
-    label = "Transcode"
-    hosts = ["hiero", "nukestudio"]
-    families = ["img", "mov"]
+            log_file.close()
 
-    def process(self, instance):
-        import os
-
-        import hiero.core.nuke as nuke
-
-        script_path = instance[0]._scriptfile
-        log_path = script_path.replace(".nk", ".log")
-        log_file = open(log_path, "w")
-        process = nuke.executeNukeScript(script_path, log_file, True)
-
-        self.poll(process)
-
-        log_file.close()
-
-        if not instance[0]._preset.properties()["keepNukeScript"]:
-            os.remove(script_path)
-            os.remove(log_path)
+            if not instance[0]._preset.properties()["keepNukeScript"]:
+                os.remove(script_path)
+                os.remove(log_path)
 
     def poll(self, process):
         import time
